+++ src/main/java/at/hannibal2/skyhanni/config/features/Garden.java
@@ -161,7 +161,7 @@ public class Garden {
     public boolean visitorColoredName = true;
 
     @Expose
-    @ConfigOption(name = "Hypixel Message", desc = "Hide the chat message from hypixel that a new visitor has arrived at your garden")
+    @ConfigOption(name = "Hypixel Message", desc = "Hide the chat message from Hypixel that a new visitor has arrived at your garden.")
     @ConfigEditorBoolean
     @ConfigAccordionId(id = 1)
     public boolean visitorHypixelArrivedMessage = true;
@@ -182,6 +182,12 @@ public class Garden {
     @ConfigEditorBoolean
     @ConfigAccordionId(id = 5)
     public boolean numberCropMilestone = true;
+
+    @Expose
+    @ConfigOption(name = "Average Milestone", desc = "Show the average crop milestone in the crop milestone inventory.")
+    @ConfigEditorBoolean
+    @ConfigAccordionId(id = 5)
+    public boolean numberAverageCropMilestone = true;
 
     @Expose
     @ConfigOption(name = "Crop Upgrades", desc = "Show the number of upgrades in the crop upgrades inventory.")